This commit was manufactured by cvs2svn to create tag 'Ext-5-1-2-rc1'.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/tags/Ext-5-1-2-rc1@10764 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/apps/snmptest.c
@@ -282,7 +282,7 @@ int
 input_variable(netsnmp_variable_list * vp)
 {
     char            buf[256];
-    int             val_len;
+    size_t          val_len;
     u_char          value[256], ch;
     oid             name[MAX_OID_LEN];
